+++ modules/auto-scaler/main.tf
@@ -1,33 +1,20 @@
-//////////////////////////////////
-// Parent Resources
-//////////////////////////////////
-
-data "azurerm_resource_group" "MAIN" {
-  name = var.config.resource_group.name
-}
+// Recommended: Make Terraform Operator a subscription owner
 
 //////////////////////////////////
 // Virtual Desktop | Auto-Scaling
 //////////////////////////////////
 
-resource "random_uuid" "MAIN" {}
-
-/*data "azuread_application" "MAIN" {
-  display_name   = "Azure Virtual Desktop"
-  #application_id = "9cdead84-a844-4324-93f2-b2e6bb768d07"
-  #object_id      = "7f6cbba6-23c8-40a6-b629-fbaddced642c"
-}*/
+data "azurerm_subscription" "CURRENT" {}
 
 data "azuread_service_principal" "MAIN" {
   display_name   = "Azure Virtual Desktop"
-  #application_id = "9cdead84-a844-4324-93f2-b2e6bb768d07"
   #object_id      = "7f6cbba6-23c8-40a6-b629-fbaddced642c"
 }
 
 resource "azurerm_role_definition" "MAIN" {
-  name        = "AzureVirtualDesktopAutoScaler"
-  description = "AzureVirtualDesktop AutoScaler Role"
-  scope       = data.azurerm_resource_group.MAIN.id
+  name        = "custom-avd-autoscale-role"
+  description = "Custom AutoScaler AVD Role"
+  scope       = data.azurerm_subscription.CURRENT.id
 
   permissions {
     actions = [
@@ -37,6 +24,11 @@ resource "azurerm_role_definition" "MAIN" {
       "Microsoft.Compute/virtualMachines/powerOff/action",
       "Microsoft.Compute/virtualMachines/start/action",
       "Microsoft.Compute/virtualMachines/read",
+      #"Microsoft.Compute/virtualMachineScaleSets/deallocate/action",
+      #"Microsoft.Compute/virtualMachineScaleSets/restart/action",
+      #"Microsoft.Compute/virtualMachineScaleSets/powerOff/action",
+      #"Microsoft.Compute/virtualMachineScaleSets/start/action",
+      #"Microsoft.Compute/virtualMachineScaleSets/read",
       "Microsoft.DesktopVirtualization/hostpools/read",
       "Microsoft.DesktopVirtualization/hostpools/write",
       "Microsoft.DesktopVirtualization/hostpools/sessionhosts/read",
@@ -44,29 +36,31 @@ resource "azurerm_role_definition" "MAIN" {
       "Microsoft.DesktopVirtualization/hostpools/sessionhosts/usersessions/delete",
       "Microsoft.DesktopVirtualization/hostpools/sessionhosts/usersessions/read",
       "Microsoft.DesktopVirtualization/hostpools/sessionhosts/usersessions/sendMessage/action",
-      "Microsoft.DesktopVirtualization/hostpools/sessionhosts/usersessions/read"
+      "Microsoft.DesktopVirtualization/hostpools/sessionhosts/usersessions/read",
     ]
-    
+
     not_actions = []
   }
   
   assignable_scopes = [
-    data.azurerm_resource_group.MAIN.id,
+    data.azurerm_subscription.CURRENT.id,
   ]
 }
 
+#resource "random_uuid" "MAIN" {}
+
 resource "azurerm_role_assignment" "MAIN" {
-  name                             = random_uuid.MAIN.result
-  scope                            = data.azurerm_resource_group.MAIN.id
+  #name                             = random_uuid.MAIN.result //Auto-generate
+  scope                            = data.azurerm_subscription.CURRENT.id
   role_definition_id               = azurerm_role_definition.MAIN.role_definition_resource_id
   principal_id                     = data.azuread_service_principal.MAIN.id
-  skip_service_principal_aad_check = true
+  skip_service_principal_aad_check = false
 }
 
 resource "azurerm_virtual_desktop_scaling_plan" "MAIN" {
-  name          = "DefaultScalingPlan"
-  friendly_name = "AzureVirtualDesktop Scaling Plan"
-  description   = "Example Azure Virtual Desktop Scaling Plan"
+  name          = "ScalingPlan"
+  friendly_name = "Scaling Plan"
+  description   = "Azure Virtual Desktop Scaling Plan"
   time_zone     = "W. Europe Standard Time"
 
   host_pool {
@@ -79,7 +73,7 @@ resource "azurerm_virtual_desktop_scaling_plan" "MAIN" {
     days_of_week                         = ["Monday", "Tuesday", "Wednesday", "Thursday", "Friday"]
     ramp_up_start_time                   = "06:00"
     ramp_up_load_balancing_algorithm     = "BreadthFirst"
-    ramp_up_minimum_hosts_percent        = 20
+    ramp_up_minimum_hosts_percent        = 40
     ramp_up_capacity_threshold_percent   = 20
     peak_start_time                      = "08:30"
     peak_load_balancing_algorithm        = "BreadthFirst"
@@ -95,6 +89,6 @@ resource "azurerm_virtual_desktop_scaling_plan" "MAIN" {
     off_peak_load_balancing_algorithm    = "DepthFirst"
   }
   
-  location            = data.azurerm_resource_group.MAIN.location
-  resource_group_name = data.azurerm_resource_group.MAIN.name
+  location            = var.config.resource_group.location
+  resource_group_name = var.config.resource_group.name
 }