+++ modules/auto-scaler/variables.tf
@@ -1,8 +1,9 @@
 variable "config" {
   type = object({
     resource_group = object({
-      name = string
+      name     = string
       location = string
+      id       = string
     })
 
     host_pool = object({