+++ modules/azure-ad/outputs.tf
@@ -1,9 +1,21 @@
-output "demo_credentials" {
+output "demo_users" {
   sensitive = true
   
-  value = {
-    url      = "https://client.wvd.microsoft.com/arm/webclient/"
-    username = azuread_user.DEMO_USER.user_principal_name
-    password = azuread_user.DEMO_USER.password
-  }
+  value = flatten([
+    for user in azuread_user.DEMO_USER: {
+      username = user.user_principal_name
+      password = user.password
+    }
+  ])
+}
+
+output "demo_admins" {
+  sensitive = true
+  
+  value = flatten([
+    for user in azuread_user.DEMO_ADMIN: {
+      username = user.user_principal_name
+      password = user.password
+    }
+  ])
 }