+++ modules/rd-app/main.tf
@@ -1,28 +1,18 @@
 //////////////////////////////////
 // Virtual Desktop Application
 //////////////////////////////////
-
-resource "azurerm_virtual_desktop_application_group" "MAIN" {
-  name          = "ExampleApps"
-  type          = "RemoteApp"
-  host_pool_id  = azurerm_virtual_desktop_host_pool.MAIN.id
-  friendly_name = "Examples"
-  description   = "Example RemoteApp Applications"
-
-  location            = azurerm_resource_group.MAIN.location
-  resource_group_name = azurerm_resource_group.MAIN.name
-}
-
+// https://registry.terraform.io/providers/hashicorp/azurerm/latest/docs/resources/virtual_desktop_application
 
 resource "azurerm_virtual_desktop_application" "MAIN" {
-  name                         = "notepad"
-  friendly_name                = "Notepad"
-  description                  = "Notepad text-editor"
-  path                         = "C:\\WINDOWS\\System32\\notepad.exe"
-  command_line_argument_policy = "DoNotAllow"
-  show_in_portal               = false
-  icon_path                    = "C:\\WINDOWS\\System32\\notepad.exe"
-  icon_index                   = 0
-  application_group_id         = azurerm_virtual_desktop_application_group.MAIN.id
+  name                         = var.config.name
+  friendly_name                = var.config.friendly_name
+  description                  = var.config.description
+  path                         = var.config.path
+  icon_path                    = var.config.icon_path
+  icon_index                   = var.config.icon_index
+  command_line_argument_policy = var.config.cli_policy
+  command_line_arguments       = var.config.cli_arguments
+  show_in_portal               = var.config.show_in_portal
+  application_group_id         = var.application_group.id
 }