+++ modules/rd-app/variables.tf
@@ -1,16 +1,24 @@
+//////////////////////////////////
+// Required
+//////////////////////////////////
+
+variable "application_group" {
+  type = object({
+    id = string
+  })
+}
+
 variable "config" {
   type = object({
-    resource_group = object({
-      name     = string
-      location = string
-      id       = string
-    })
+    name      = string
+    path      = string
+    icon_path = string
+    
+    friendly_name  = optional(string)
+    description    = optional(string)
+    icon_index     = optional(number, 0)
+    show_in_portal = optional(bool, false)
+    cli_policy     = optional(string, "DoNotAllow")
+    cli_arguments  = optional(string)
   })
-
 }
-
-variable "aad_desktop_users" {
-  type    = set(string)
-  default = []
-}
-