+++ outputs.tf
@@ -1,4 +1,14 @@
 output "login" {
   sensitive = true
-  value     = one(module.USER_ACCESS.*.demo_credentials)
+  value = flatten([
+    concat(
+      module.USER_ACCESS.*.demo_users,
+      module.USER_ACCESS.*.demo_admins,
+    )
+  ])
+}
+
+output "azure_url" {
+  sensitive = false
+  value     = "https://client.wvd.microsoft.com/arm/webclient/"
 }