+++ CHANGELOG.md
@@ -1,3 +1,16 @@
-# Unreleased (2022-10-11)
+# CHANGELOG
 
-* Initial version created
+## v2022-10-13
+#### Info
+  * Cleanup unused modules
+  * Set default session-host image to 'win11-22h2-avd-m365'
+#### Features
+  * NSG-rules module (fqdn not supported)
+  * Auto-scaling module
+  * RemoteApps module
+  * Session-hosts module
+  * AAD users module (incl. demo users/admins)
+---
+## v2022-10-11
+#### Info
+  * Initial version